refactor(core): extract firebase-scrypt into separate module
